More frontend renaming and cleanup
