+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adaptors/all_room_adaptors/ListRoomsViewController.java
@@ -1,4 +1,4 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.game_room_adaptors;
+package dev.totallyspies.spydle.frontend.interface_adaptors.all_room_adaptors;
 
 import dev.totallyspies.spydle.frontend.interface_adaptors.game_view_adaptors.SwitchViewEvent;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,14 +6,16 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 @Component
-public class GameRoomController {
+public class ListRoomsViewController {
+
     @Autowired
     private ApplicationEventPublisher publisher;
 
     /*
     Method called when View All Rooms Button is Pressed
      */
-    public void changeView(String view) {
-        publisher.publishEvent(new SwitchViewEvent(this, view));
+    public void openWelcomeView() {
+        publisher.publishEvent(new SwitchViewEvent(this, "WelcomeView"));
     }
+
 }
