+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adaptors/all_room_adaptors/ListRoomsViewModel.java
@@ -3,5 +3,6 @@ package dev.totallyspies.spydle.frontend.interface_adaptors.all_room_adaptors;
 import org.springframework.stereotype.Component;
 
 @Component
-public class AllRoomViewModel {
+public class ListRoomsViewModel {
+
 }
