+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adaptors/game_room_adaptors/GameRoomViewModel.java
@@ -4,4 +4,5 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class GameRoomViewModel {
+
 }
