+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adaptors/welcome_screen_adaptors/WelcomeViewController.java
@@ -1,4 +1,4 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.game_over_adaptors;
+package dev.totallyspies.spydle.frontend.interface_adaptors.welcome_screen_adaptors;
 
 import dev.totallyspies.spydle.frontend.interface_adaptors.game_view_adaptors.SwitchViewEvent;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,14 +6,16 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 @Component
-public class GameOverController {
+public class WelcomeViewController {
+
     @Autowired
     private ApplicationEventPublisher publisher;
 
     /*
     Method called when View All Rooms Button is Pressed
      */
-    public void changeView(String view) {
-        publisher.publishEvent(new SwitchViewEvent(this, view));
+    public void openListRoomsView() {
+        publisher.publishEvent(new SwitchViewEvent(this, "ListRoomsView"));
     }
+
 }
