+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adaptors/welcome_screen_adaptors/WelcomeViewModel.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Getter
-public class WelcomeScreenViewModel {
+public class WelcomeViewModel {
 
     private String playerName;
     private String roomCode;
