+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adaptors/welcome_screen_adaptors/WelcomeViewPresenter.java
@@ -3,7 +3,6 @@ package dev.totallyspies.spydle.frontend.interface_adaptors.welcome_screen_adapt
 import org.springframework.stereotype.Component;
 
 @Component
-public class WelcomeScreenPresenter {
-
+public class WelcomeViewPresenter {
 
 }
