+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/use_cases/JoinGameInteractor.java
@@ -1,4 +1,5 @@
 package dev.totallyspies.spydle.frontend.use_cases;
 
 public class JoinGameInteractor {
+
 }
