+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/views/GameOverView.java
@@ -1,6 +1,6 @@
 package dev.totallyspies.spydle.frontend.views;
 
-import dev.totallyspies.spydle.frontend.interface_adaptors.game_over_adaptors.GameOverController;
+import dev.totallyspies.spydle.frontend.interface_adaptors.game_over_adaptors.GameOverViewController;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.swing.*;
@@ -10,8 +10,9 @@ import java.awt.event.ActionListener;
 
 @org.springframework.stereotype.Component
 public class GameOverView extends JPanel {
+
     @Autowired
-    private GameOverController controller;
+    private GameOverViewController controller;
 
     public GameOverView() {
 
@@ -67,7 +68,7 @@ public class GameOverView extends JPanel {
         backButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                controller.changeView("WelcomeScreenView"); // Open the rooms page (AllRoomScreen.AllRoomsPage)
+                controller.openWelcomeView(); // Open the rooms page (AllRoomScreen.AllRoomsPage)
             }
         });
 
@@ -126,4 +127,5 @@ public class GameOverView extends JPanel {
             frame.setVisible(true);
         });
     }
+
 }
